Moved NCCL communicator wrapper to helpers
